Rename files, delete old files
