+++ utils/setup.py
@@ -6,7 +6,7 @@ import coloredlogs
 from openapi_core import create_spec
 import openapi_core.schema.servers.models
 import yaml
-import config.hosts
+import config.config
 
 
 def setup():
@@ -15,10 +15,10 @@ def setup():
     parser = argparse.ArgumentParser()
     parser.add_argument('host', type=str, nargs='?', default='local')
     c_args = parser.parse_args()
-    host = config.hosts.HOSTS.get(c_args.host, config.hosts.HOSTS.get('local'))
+    host = config.config.HOSTS.get(c_args.host, config.config.HOSTS.get('local'))
     logging.info('Testing host %s', host)
 
-    spec = parse_spec(config.hosts.SPEC)
+    spec = parse_spec(config.config.SPEC)
     server = openapi_core.schema.servers.models.Server(host)
     spec.servers.append(server)
     return host, spec
